refacto(Profile): create SocialNetwork subcomponent
